app/requirements.txt
alembic==1.3.2
beautifulsoup4==4.8.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.0
cryptography==2.8
flask-cors==3.0.8
flask-marshmallow==0.10.1
flask-migrate==2.5.2
flask-sqlalchemy==2.4.1
flask==1.1.1
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
google==2.0.3
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
mako==1.1.0
markupsafe==1.1.1
marshmallow-sqlalchemy==0.21.0
marshmallow==3.3.0
mysqlclient==1.4.6
oauthlib==3.1.0
protobuf==3.11.2
ptvsd==4.3.2
pyasn1-modules==0.2.7
pyasn1==0.4.8
pycparser==2.19
pyjwt==1.7.1
pymysql==0.9.3
pyopenssl==19.1.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlalchemy==1.3.12
urllib3==1.25.7
werkzeug==0.16.0
gunicorn==20.0.4
